move color related tools into lib directory
